primitives分割 Phase 2: compare.rs を切り出す
EQ/NEQ/LT/GT/LE/GE の実装を src/primitives/compare.rs に移動し、
primitives.rs をファサードとして更新する。
EQ/NEQ は resolve_str_cell 依存を削除し、Cell の PartialEq に委譲する。
